+++ src/components/Dashboard/ChangePasswordSuggested.tsx
@@ -1,18 +1,18 @@
 import { CopyToClipboard } from "components/Common/CopyToClipboard";
 import { NewPasswordForm } from "components/Common/NewPasswordForm";
 import { useAppSelector } from "eduid-hooks";
-import { useRef } from "react";
+import React, { useRef } from "react";
 import { FormattedMessage } from "react-intl";
-import { ChangePasswordChildFormProps } from "./ChangePasswordForm";
+import { ChangePasswordChildFormProps } from "./ChangePassword";
 
 export default function ChangePasswordSuggestedForm(props: ChangePasswordChildFormProps) {
   const ref = useRef<HTMLInputElement>(null);
   const suggested_password = useAppSelector((state) => state.chpass.suggested_password);
   // Form field validator
   const required = (value: string) => (value ? undefined : "required");
-  console.log("props", props);
+
   return (
-    <>
+    <React.Fragment>
       <div className="reset-password-input">
         <label htmlFor="copy-new-password">
           <FormattedMessage defaultMessage="New password" description="Set new password" />
@@ -21,7 +21,7 @@ export default function ChangePasswordSuggestedForm(props: ChangePasswordChildFo
           name="copy-new-password"
           id="copy-new-password"
           ref={ref}
-          defaultValue={suggested_password ? suggested_password : ""}
+          defaultValue={suggested_password}
           readOnly={true}
         />
         <CopyToClipboard ref={ref} />
@@ -29,26 +29,9 @@ export default function ChangePasswordSuggestedForm(props: ChangePasswordChildFo
       <NewPasswordForm
         suggested_password={suggested_password}
         submitNewPasswordForm={props.formProps.handleSubmit}
-        submitButtonText={
-          <FormattedMessage defaultMessage="accept password" description="Set new password (accept button)" />
-        }
+        submitButtonText={<FormattedMessage defaultMessage="Ok" description="Set new password (ok button)" />}
         handleCancel={props.handleCancel}
       />
-    </>
-    // <form id="passwordsview-form" role="form" onSubmit={props.formProps.handleSubmit}>
-    //   <fieldset>
-    //     <FinalField<string>
-    //       name="suggested"
-    //       component={TextInput}
-    //       componentClass="input"
-    //       type="text"
-    //       id="suggested-password-field"
-    //       className="suggested-password"
-    //       label={<FormattedMessage defaultMessage="Suggested password" description="chpass suggested password" />}
-    //       disabled={true}
-    //       autoComplete="new-password"
-    //     />
-    //   </fieldset>
-    // </form>
+    </React.Fragment>
   );
 }
